separated socket server into own file
